refactor(parser): extract inline parsing helpers
