+++ packages/markdown-parser/src/parser/inline-parsers/emphasis-parser.ts
@@ -1,5 +1,6 @@
 import type { EmphasisNode, MarkdownToken, ParsedNode, ParseOptions } from '../../types'
 import type { ParseInlineTokensFn } from './inline-parser-types'
+import { collectDelimitedInlineTokens } from './token-range'
 
 export function parseEmphasisToken(
   tokens: MarkdownToken[],
@@ -11,17 +12,13 @@ export function parseEmphasisToken(
   nextIndex: number
 } {
   const children: ParsedNode[] = []
-  let emText = ''
-  let i = startIndex + 1
-  const innerTokens: MarkdownToken[] = []
-
-  // Process tokens between em_open and em_close
-  while (i < tokens.length && tokens[i].type !== 'em_close') {
-    const tokenText = tokens[i] as MarkdownToken & { text?: unknown }
-    emText += String(tokens[i].content ?? tokenText.text ?? '')
-    innerTokens.push(tokens[i])
-    i++
-  }
+  const { content: emText, innerTokens, nextIndex } = collectDelimitedInlineTokens(
+    tokens,
+    startIndex,
+    'em_close',
+    undefined,
+    true,
+  )
 
   // Parse inner tokens to handle nested elements
   children.push(...parseInlineTokens(innerTokens, undefined, undefined, options))
@@ -31,9 +28,6 @@ export function parseEmphasisToken(
     children,
     raw: `*${emText}*`,
   }
-
-  // Skip to after em_close
-  const nextIndex = i < tokens.length ? i + 1 : tokens.length
 
   return { node, nextIndex }
 }
