+++ packages/markdown-parser/src/parser/inline-parsers/insert-parser.ts
@@ -1,5 +1,6 @@
 import type { InsertNode, MarkdownToken, ParsedNode, ParseOptions } from '../../types'
 import type { ParseInlineTokensFn } from './inline-parser-types'
+import { collectDelimitedInlineTokens } from './token-range'
 
 export function parseInsertToken(
   tokens: MarkdownToken[],
@@ -11,16 +12,7 @@ export function parseInsertToken(
   nextIndex: number
 } {
   const children: ParsedNode[] = []
-  let insText = ''
-  let i = startIndex + 1
-  const innerTokens: MarkdownToken[] = []
-
-  // Process tokens between ins_open and ins_close
-  while (i < tokens.length && tokens[i].type !== 'ins_close') {
-    insText += String(tokens[i].content ?? '')
-    innerTokens.push(tokens[i])
-    i++
-  }
+  const { content: insText, innerTokens, nextIndex } = collectDelimitedInlineTokens(tokens, startIndex, 'ins_close')
 
   // Parse inner tokens to handle nested elements
   children.push(...parseInlineTokens(innerTokens, undefined, undefined, options))
@@ -30,9 +22,6 @@ export function parseInsertToken(
     children,
     raw: `++${String(insText)}++`,
   }
-
-  // Skip to after ins_close
-  const nextIndex = i < tokens.length ? i + 1 : tokens.length
 
   return { node, nextIndex }
 }
